renamed dialogs_ui package
